requirements.txt
flask
dnspython
requests
beautifulsoup4
playwright
lxml
python-dotenv